requirements.txt
absl-py==2.1.0
accelerate==1.2.1
acres==0.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==4.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==5.0.1
blinker==1.8.2
blis==0.7.11
cachetools==5.2.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
commonmark==0.9.1
confection==0.1.5
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.10
datasets==3.2.0
debugpy==1.6.3
decorator==5.1.1
dill==0.3.8
distlib==0.3.4
distro==1.9.0
dnspython==2.7.0
docutils==0.19
entrypoints==0.4
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==0.10.0
filelock==3.16.1
fitz==0.0.1.dev2
Flask==2.0.3
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.9
GitPython==3.1.27
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
h5py==3.8.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
huggingface-hub==0.26.1
idna==3.10
importlib-metadata==4.12.0
importlib_resources==6.4.5
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.1
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonschema==4.13.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.2
keyring==23.8.2
kiwisolver==1.4.8
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
looseversion==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.5.5
networkx==3.4.2
nh3==0.2.20
nibabel==5.0.0
nipype==1.8.3
nltk==3.9.1
numpy==1.24.4
openai==1.55.3
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==24.1
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
pickleshare==0.7.5
pillow==11.0.0
pipdeptree==2.23.4
pkginfo==1.12.0
platformdirs==2.5.1
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.3
prov==2.0.1
psutil==5.9.1
psycopg2==2.9.10
pure-eval==0.2.2
puremagic==1.28
pyarrow==18.1.0
pycparser==2.22
pydantic==1.10.12
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.2
Pygments==2.18.0
pymongo==4.4.1
Pympler==1.0.1
PyMuPDF==1.24.14
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
pywin32==308
pywin32-ctypes==0.2.3
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
readme_renderer==43.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==13.9.4
safetensors==0.4.5
scikit-fuzzy==0.5.0
scikit-learn==1.2.2
scikit-surprise==1.1.4
scipy==1.10.1
selenium==4.26.1
semver==3.0.2
sentence-transformers==3.3.1
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.5
smmap==5.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
srsly==2.4.8
stack-data==0.4.0
streamlit==1.12.0
streamlit-cookies-manager==0.2.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
sympy==1.13.1
termcolor==2.5.0
textblob==0.18.0.post0
thinc==8.1.8
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.0
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.66.6
traitlets==5.14.3
traits==6.4.3
transformers==4.41.0
trio==0.27.0
trio-websocket==0.11.1
twine==4.0.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.2
urllib3==2.2.3
validators==0.20.0
virtualenv==20.14.1
wasabi==1.1.3
watchdog==2.1.9
wcwidth==0.2.5
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.0.3
widgetsnbextension==4.0.2
wincertstore==0.2
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
